Prepare version for 0.3.10 release
Abolish slurping of the version file. Unix convention demand files have
a final newline. When I edit VERSION, it will end up with such a newline
and this newline will find its way into the version string, making it
malformed. Therefore provide the version in the usual way. What was the
slurping good for, anyway?